// src/app/enums/modifier-discriminator.enum.ts
export enum ModifierDiscriminator {
    AbilityScore_Change = 0,
    Defense_Add = 1,
    HitPointMaximum_Change = 2,
    Influence_Condition_Add = 3,
    Influence_Custom_Add = 4,
    Influence_DamageType_Add = 5,
    Proficiency_Add = 6,
    Sense_Add = 7,
    Skill_Proficient = 8,
    Speed_Add = 9,
    Spell_Add = 10
}